modified:   codelite/joies/goose.mk 	modified:   codelite/joies/jo_core.mk 	modified:   codelite/joies/jo_pcap.mk 	modified:   codelite/joies/jo_utils.mk 	modified:   codelite/joies/joies.mk 	modified:   codelite/joies/joies.tags 	modified:   codelite/joies/joies.workspace.session 	modified:   codelite/joiss/AipstarHost.mk 	modified:   codelite/joiss/AironixHost.mk 	modified:   codelite/joiss/DahuaHost.mk 	modified:   codelite/joiss/DahuaHost.project 	modified:   codelite/joiss/HikHost.mk 	modified:   codelite/joiss/HttpRequest.mk 	modified:   codelite/joiss/JofsReader.mk 	modified:   codelite/joiss/JoissManager.mk 	modified:   codelite/joiss/JosStream.mk 	modified:   codelite/joiss/JospControl.mk 	modified:   codelite/joiss/JospRequest.mk 	modified:   codelite/joiss/LocalConfig.mk 	modified:   codelite/joiss/MCConfig.mk 	modified:   codelite/joiss/MCControl.mk 	modified:   codelite/joiss/OnvifHost.mk 	modified:   codelite/joiss/PSStream.mk 	modified:   codelite/joiss/RecordManager.mk 	modified:   codelite/joiss/RtpStream.mk 	modified:   codelite/joiss/RtspRequest.mk 	modified:   codelite/joiss/SamsungHost.mk 	modified:   codelite/joiss/SonyHost.mk 	modified:   codelite/joiss/StreamManager.mk 	modified:   codelite/joiss/TSStream.mk 	modified:   codelite/joiss/core.mk 	modified:   codelite/joiss/joiss.mk 	modified:   codelite/joiss/joiss.tags 	modified:   codelite/joiss/joiss.workspace.session 	modified:   codelite/joiss/misc.mk 	modified:   codelite/joiss/utils.mk 	modified:   libs/dahua/libdhdvr.so modified:   libs/dahua/libdhnetsdk.so 	deleted:    libs/dahua/libdhparser.so 	modified:   scripts/joiss 	modified:   scripts/joiss-deamon 	modified:   src/control/record/RecoderManager.cpp 	modified:   src/control/record/RecordMediaObj.cpp 	modified:   src/core/j_type.h 	modified:   src/goose/x_goose.cpp 	modified:   src/host/dahua/DahuaAdapter.h 	modified:   src/host/dahua/DahuaChannel.cpp 	modified:   src/host/dahua/DahuaChannel.h 	modified:   src/host/dahua/DahuaParser.cpp 	modified:   src/host/dahua/DahuaParser.h 	modified:   src/host/dahua/DahuaStream.h 	modified:   src/host/dahua/dhnetsdk.h

+++ src/host/dahua/DahuaAdapter.h
@@ -36,7 +36,7 @@ private:
 	char *GetRemotePw() const { return (char *)m_password; }
 	LONG GetDevID() const { return m_devID; }
 
-	static void OnConnectCallBack(LONG hHandle, char *pDVRIP, LONG nDVRPort, DWORD dwUser)
+	static void OnConnectCallBack(long hHandle, char *pDVRIP, int nDVRPort, long dwUser)
 	{
 		CDahuaAdapter *pThis = reinterpret_cast<CDahuaAdapter *>(dwUser);
 		if (pThis != NULL)
